Grid and Cells defined
Cells for GameOfLife have been defined to display the appropriate text
based off of the Cell's living status. Cells can be parsed from the
AliveCellString or the DeadCellString. Grids have been defined as a 2
dimensional array of Cells where the number of rows and columns must be
specified in the constructor. The Grid ToString will display the grid as

4 8
........
....*...
...**...
........

where "4" is the height of the Grid, "8" is the width of the Grid, "."
represents a dead Cell, and "*" represents a live Cell.