CellEnvironments contain neighbours of a Cell
CellEnvironments are created by providing the coordinates (row, column)
and Grid of a Cell. The environment will collect the surrounding Cells
of the location provided based off of the rules defined in the extension
of CellEnvironment. EdgedEnvironment stops at the edge of the Grid so
any position that would be outside of the Grid's bounds is not collected
as a neighbouring Cell.