Console will display game
The GameOfLife console can load a game from a file whose path is
provided as the first argument to the program. A second argument of
true/false can be provided as to if the Grid edges should wrap. The
console will progress through generations as "Enter" is hit until "Q" or
"q" is entered to quit.